steamcompmgr: write WM_STATE from a long array
Xlib reads a long per element for a format-32 property, so a uint32_t pair
fed the icon-window field from past the end of the array on LP64.

Patch-Name: wm-state-long

+++ src/steamcompmgr.cpp
@@ -4442,7 +4442,7 @@ found:;
 
 static void set_wm_state( xwayland_ctx_t *ctx, Window win, uint32_t state )
 {
-	uint32_t wmState[] = { state, None };
+	long wmState[] = { long( state ), None };
 	XChangeProperty(ctx->dpy, win, ctx->atoms.WMStateAtom, ctx->atoms.WMStateAtom, 32,
 				PropModeReplace, (unsigned char *)wmState,
 				sizeof(wmState) / sizeof(wmState[0]));
